requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.0.0
mcp>=1.0.0
rich>=13.7.0
textual>=0.47.1
pyyaml>=6.0.1

# CLI and utilities
click>=8.1.7
python-dotenv>=1.0.0
prompt_toolkit>=3.0.43

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Development dependencies (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Logging and monitoring
colorlog>=6.8.0
